zerotier-one: Fix libnatpmp build
The zerotier-one build required libnatpmp, but it based this on /usr/lib
rather than the local cross-compilation environment.

The make-linux.mk Makefile is fairly unconventional with limited
configuration macros.  But a particular combination of macro and
environment variable configuration allows us to both build the local
exp/libnatbib source, and to integrate it into the zerotier-one build.