yabasanshiro-sa: n2.cmake (retro arena) as include
The yabasanshiro-sa build was struggling to get the glut/n2 libraries
compiled and/or linked correctly.  The issue went back to handling of
the n2.cmake header in retro arena.

Instead of using n2.cmake as the toolchain file, we now use the general
cmake header in toolchain/ as the toolchain file, and append n2.cmake as
an include header.